added skill list

+++ CodeInject/NPC/Monster.cs
@@ -13,8 +13,8 @@ namespace CodeInject.NPC
     {
         public ushort ID { get; set; }
         public Vector3 Position { get; set; }
-        public int NPCModelNameID { get; set; }
-
+        public ushort NPCModelNameID { get; set; }
+         
         public Monster(ushort id)
         {
             ID = id;
